修改

+++ webSSR/src/localData/LocalStorage_.ts
@@ -1,3 +1,5 @@
+import { Env } from "@/_d/Env";
+
 /**
  * 封装后的本地数据类
  * 将把会数据以json的格式保存
@@ -10,6 +12,7 @@ export class LocalStorage_ {
      * @param _f 设置前处理
      */
     static setItem(key: string, value: any, _f?: (s: string) => string) {
+        if (Env.ifS) { return; }
         //直接保存为json数据
         localStorage.setItem(key, _f ? _f(JSON.stringify(value)) : JSON.stringify(value));
     }
@@ -20,6 +23,7 @@ export class LocalStorage_ {
      * @param _f 获取前处理
      */
     static getItem<D = any>(key: string, _f?: (s: string) => string): D | null {
+        if (Env.ifS) { return null; }
         try {
             return JSON.parse(_f ? _f(localStorage.getItem(key)!) : localStorage.getItem(key)!);
         } catch {
@@ -34,6 +38,7 @@ export class LocalStorage_ {
      * @param key 名字
      */
     static removeItem(key: string) {
+        if (Env.ifS) { return; }
         localStorage.removeItem(key);
     }
 
@@ -41,6 +46,7 @@ export class LocalStorage_ {
      * 清理本地的全部数据
      */
     static clear() {
+        if (Env.ifS) { return; }
         localStorage.clear();
     }
 }
